Remove ArtifactView
